+++ packages/types-known/src/index.spec.ts
@@ -2,7 +2,7 @@
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
-import { TypeRegistry } from '../create/registry';
+import { TypeRegistry } from '@polkadot/types/create/registry';
 import { getModuleTypes } from './';
 
 const registry = new TypeRegistry();